requirements_ui.txt
aliyun-python-sdk-cdn==3.7.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-dcdn==2.2.1
aliyun-python-sdk-domain==3.14.7
annotated-types==0.5.0
APScheduler==3.9.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
bidict==0.22.1
blinker==1.6.2
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.7
coloredlogs==15.0.1
cryptography==36.0.2
decorator==5.1.1
dnspython==2.2.0
exceptiongroup==1.1.1
Flask==2.3.3
Flask-APScheduler==1.12.4
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
huaweicloudsdkcdn==3.1.32
huaweicloudsdkcore==3.1.32
huaweicloudsdkeps==3.1.32
humanfriendly==10.0
idna==3.3
importlib-metadata==6.8.0
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
json2html==1.3.0
loguru==0.6.0
loguru-caplog==0.2.0
MarkupSafe==2.1.3
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyOpenSSL==22.0.0
pyparsing==3.0.8
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.6.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.27.1
requests-futures==1.0.0
requests-toolbelt==0.10.1
retry==0.9.2
retrying-async==1.2.0
riprova==0.2.7
simplejson==3.18.4
six==1.16.0
SQLAlchemy==2.0.20
tenacity==8.0.1
tencentcloud-sdk-python==3.0.624
tencentcloud-sdk-python-cdn==3.0.624
tencentcloud-sdk-python-ecdn==3.0.624
throttler==1.2.1
timeout-decorator==0.5.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.3.7
zipp==3.16.2
